requirements.txt
torch==2.0.0
torchvision==0.15.0
opencv-python==4.10.0.84
albumentations==1.3.1
albucore==0.0.33
numpy==1.24.0
matplotlib==3.7.0
scikit-learn==1.3.0
tqdm==4.65.0
PyYAML==6.0.1
